Complete Ninja adapter rebase integration (#339)
Remove obsolete runner wrappers and migrate remaining request fixtures to
the options and stream-policy APIs. Split the focused execution-context
tests into their own module so the process test module remains within the
repository size limit.
