+++ src/runner/process/mod.rs
@@ -8,6 +8,9 @@ use std::{io, path::Path, process::Command};
 mod child_exit;
 mod command_list_telemetry;
 mod command_logging;
+
+#[cfg(test)]
+mod exit_status_tests;
 mod dyndep_files;
 mod dyndep_retention;
 mod dyndep_telemetry;
