+++ src/runner/process/tests.rs
@@ -1,7 +1,7 @@
 //! Unit and property tests for Ninja process helpers.
 
 use super::super::{NINJA_ENV, NINJA_PROGRAM};
-use super::child_exit::{check_exit_status_with_context, finalize_streaming};
+use super::child_exit::finalize_streaming;
 #[cfg(unix)]
 use super::command_list_telemetry::COMMAND_LIST_FAILURE_DURATION;
 use super::streaming::ForwardStats;
@@ -14,9 +14,10 @@ use metrics_util::{
     debugging::{DebugValue, DebuggingRecorder},
 };
 use mockable::MockEnv;
-use monotony::test_util::FixedMonotonicClock;
 #[cfg(unix)]
 use monotony::StdMonotonicClock;
+#[cfg(unix)]
+use monotony::test_util::FixedMonotonicClock;
 use proptest::prelude::*;
 use rstest::{fixture, rstest};
 use std::ffi::OsString;
@@ -25,9 +26,9 @@ use std::path::Path;
 use std::path::PathBuf;
 #[cfg(unix)]
 use std::process::Stdio;
-use std::process::ExitStatus;
 use std::thread;
-use std::time::{Duration, Instant};
+#[cfg(unix)]
+use std::time::Duration;
 use tracing_subscriber::filter::LevelFilter;
 
 /// Open a capability directory rooted at an owned UTF-8 temporary directory.
@@ -99,148 +100,6 @@ fn resolve_ninja_program_with_converts_the_resolved_path(
 ) {
     let resolved = resolve_ninja_program_with(&ninja_env);
     assert_eq!(resolved, std::path::PathBuf::from("/opt/ninja"));
-}
-
-#[cfg(unix)]
-fn exit_status(code: i32) -> ExitStatus {
-    use std::os::unix::process::ExitStatusExt;
-
-    ExitStatus::from_raw(code << 8)
-}
-
-#[cfg(windows)]
-fn exit_status(code: i32) -> ExitStatus {
-    use std::os::windows::process::ExitStatusExt;
-
-    ExitStatus::from_raw(code as u32)
-}
-
-fn command_log_context() -> CommandLogContext {
-    CommandLogContext::from_command(&Command::new("ninja"))
-}
-
-fn captured_exit_result(status: ExitStatus) -> (io::Result<()>, Vec<String>) {
-    let clock = FixedMonotonicClock::with_elapsed(Duration::ZERO);
-    let context = command_log_context();
-    let failure_context = ExitFailureContext {
-        operation: "clean",
-        stderr_mode: StderrMode::Suppress,
-        command_list_failure: None,
-        clock: &clock,
-        started_at: Instant::now(),
-    };
-    with_test_subscriber(LevelFilter::WARN, |captured| {
-        let result = check_exit_status_with_context(status, &context, &failure_context);
-        (result, captured.snapshot())
-    })
-}
-
-#[test]
-fn successful_ninja_exit_returns_ok_without_exit_failure_event() {
-    let (result, events) = captured_exit_result(exit_status(0));
-
-    assert!(result.is_ok(), "a successful Ninja exit should succeed");
-    assert!(
-        events.is_empty(),
-        "a successful Ninja exit should emit no exit-failure warning: {events:?}"
-    );
-}
-
-#[test]
-fn failed_ninja_exit_records_exit_status_diagnostics() {
-    let (result, events) = captured_exit_result(exit_status(1));
-
-    assert!(result.is_err(), "a failed Ninja exit should return an error");
-    let [event] = events.as_slice() else {
-        panic!("expected one exit-failure event, got {events:?}");
-    };
-    assert!(
-        event.contains("operation=\"clean\""),
-        "exit failure should retain the operation label: {event}"
-    );
-    assert!(
-        event.contains("failure_category=\"exit_status\""),
-        "exit failure should identify its category: {event}"
-    );
-    assert!(
-        event.contains("status="),
-        "exit failure should include the child status: {event}"
-    );
-}
-
-#[cfg(unix)]
-fn fake_ninja_program() -> anyhow::Result<(tempfile::TempDir, PathBuf)> {
-    use cap_std::{
-        ambient_authority,
-        fs::{Dir, PermissionsExt},
-    };
-
-    let temp_dir = tempfile::tempdir()?;
-    let directory = Dir::open_ambient_dir(temp_dir.path(), ambient_authority())?;
-    directory.write("ninja", "#!/bin/sh\nexit 0\n")?;
-    let mut permissions = directory.metadata("ninja")?.permissions();
-    permissions.set_mode(0o700);
-    directory.set_permissions("ninja", permissions)?;
-    let program = temp_dir.path().join("ninja");
-    Ok((temp_dir, program))
-}
-
-#[cfg(unix)]
-#[test]
-fn build_and_tool_execution_preserve_operation_labels() -> anyhow::Result<()> {
-    let (_program_dir, program) = fake_ninja_program()?;
-    let build_file = tempfile::NamedTempFile::new()?;
-    let options = NinjaProcessOptions::default();
-    let env = CommandEnv::inherit();
-    let target_names = vec![String::from("default")];
-    let targets = BuildTargets::new(&target_names);
-    let clock = FixedMonotonicClock::with_elapsed(Duration::ZERO);
-    let (build_result, build_events) = with_test_subscriber(LevelFilter::INFO, |captured| {
-        let result = run_ninja_build_internal(
-            NinjaBuildRequest {
-                program: &program,
-                options: &options,
-                build_file: build_file.path(),
-                targets: &targets,
-                env: &env,
-                stderr_mode: StderrMode::Forward,
-            },
-            None,
-            &clock,
-        );
-        (result, captured.snapshot())
-    });
-    build_result?;
-    anyhow::ensure!(
-        build_events
-            .iter()
-            .any(|event| event.contains("operation=\"build\"")),
-        "build execution should log the build operation: {build_events:?}"
-    );
-
-    let (tool_result, tool_events) = with_test_subscriber(LevelFilter::INFO, |captured| {
-        let result = run_ninja_tool_internal(
-            NinjaToolRequest {
-                program: &program,
-                options: &options,
-                build_file: build_file.path(),
-                tool: "clean",
-                env: &env,
-                stderr_mode: StderrMode::Forward,
-            },
-            None,
-            &clock,
-        );
-        (result, captured.snapshot())
-    });
-    tool_result?;
-    anyhow::ensure!(
-        tool_events
-            .iter()
-            .any(|event| event.contains("operation=\"clean\"")),
-        "tool execution should log the request tool: {tool_events:?}"
-    );
-    Ok(())
 }
 
 // `proptest!` owns the generated function signature, so rstest cannot inject
